# doc/api.yaml
openapi: 3.0.1
info:
  title: mcs api
  description: list of main api
  version: 1.0.0
servers:
- url: https://mcs.cloud273.com/app/api
paths:
...
  /admin/clinic/list:
    get:
      tags:
      - admin/clinic
      summary: Get list of clinic
      operationId: getListClinicByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
...
  /admin/appointment/list:
    get:
      tags:
      - admin/appointment
      summary: Get list of appointment
      operationId: getListAppointmentByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: statusTypes
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "W2FjY2VwdGVkXQ=="
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-25T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /admin/appointment/reject:
    patch:
      tags:
      - admin/appointment
      summary: Reject an appointment
      operationId: rejectAppointmentByAdmin
      parameters:
      - in: header
        name: os
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  $ref: '#/components/schemas/ID'
                note:
                  type: string
                  example: "Rejected by admin"
                  default: "Rejected by admin"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        406:
          description: Can not be rejected
        500:
          description: Unknown
...
  /clinic/appointment:
    get:
      tags:
      - clinic/appointment
      summary: Get appointment
      operationId: getAppointmentByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: id
        in: query
        description: id
        required: true
        schema:
          type: number
          default: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentDetail'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
  /clinic/appointment/list:
    get:
      tags:
      - clinic/appointment
      summary: Get list of appointment
      operationId: getListAppointmentByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: statusTypes
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "W2FjY2VwdGVkXQ=="
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-25T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /clinic/appointment/accept:
    patch:
      tags:
      - clinic/appointment
      summary: Accept an appointment
      operationId: acceptAppointmentByClinic
      parameters:
      - in: header
        name: os
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  $ref: '#/components/schemas/ID'
                note:
                  type: string
                  example: "Rejected by clinic"
                  default: "Rejected by clinic"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        406:
          description: Can not be accepted
        500:
          description: Unknown 
  /clinic/appointment/reject:
    patch:
      tags:
      - clinic/appointment
      summary: Reject an appointment
      operationId: rejectAppointmentByClinic
      parameters:
      - in: header
        name: os
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  $ref: '#/components/schemas/ID'
                note:
                  type: string
                  example: "Rejected by clinic"
                  default: "Rejected by clinic"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        406:
          description: Can not be rejected
        500:
          description: Unknown
...
  /doctor/appointment/list:
    get:
      tags:
      - doctor/appointment
      summary: Get list of appointment
      operationId: getListAppointmentByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: statusTypes
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "W2FjY2VwdGVkXQ=="
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-25T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /patient/appointment/list:
    get:
      tags:
      - patient/appointment
      summary: Get list of appointment
      operationId: getListAppointmentByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: statusTypes
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "W2FjY2VwdGVkXQ=="
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-25T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
components:
  schemas:
...
    CurrencyUnit:
      type: string
      example: vnd
      default: vnd
      enum: [vnd, usd]
...
    DateTime:
      type: string
      format: yyyy-MM-ddTHH:mm:ssZ
      example: "2020-01-01T08:00:00+07:00"
      default: "2020-01-01T08:00:00+07:00"
...
    ID:
      type: integer
      format: int64
      example: 1
      default: 1
...
    MedicationType:
      type: string
      example: highBP
      default: highBP
      enum: [highBP, highCholesterol, pregnant, cancer]
...
    Description:
      maxLength: 1000
      type: string
      example: "Description"
      default: "Description"
    Number:
      type: number
      example: 10
      default: 10
...
    PackageType:
      type: string
      example: "classic"
      default: "classic"
      enum: ["classic", "telemed"]
...
    SpecialtyCode:
      type: string
      example: "primary"
      default: "primary"
      enum: ["primary", "cardiology"]
...
    Trilian:
      type: string
      example: yes
      default: yes
      enum: [yes, no, unknown]
...
    VisitTime:
      type: number
      example: 600
      default: 600
...
    SuccessMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: successful
          default: successful
...
    Price:
      type: object
      required: [amount, currency]
      properties:
        amount:
          $ref: '#/components/schemas/Number'
        currency:
          $ref: '#/components/schemas/CurrencyUnit'
...
    AllergyCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Allergy"
          default: "Allergy"
        description:
          $ref: '#/components/schemas/Description'
...
    AppointmentCreate:
      type: object
      required: [name]
      properties:
        begin:
          $ref: '#/components/schemas/DateTime'
        specialty:
          $ref: '#/components/schemas/SpecialtyCode'
        type:
          $ref: '#/components/schemas/PackageType'
        price:
          $ref: '#/components/schemas/Price'
        visitTime:
          $ref: '#/components/schemas/VisitTime'
        symptoms:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/SymptomCreate'
        allergies:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/AllergyCreate'
        surgeries:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/SurgeryCreate'
        medications:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/MedicationCreate'
        packageId:
          $ref: '#/components/schemas/ID'
    AppointmentBasic:
      allOf:
        - $ref: '#/components/schemas/AppointmentCreate'
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Number'
        code:
          type: string
          example: "10567345"
          default: "10567345"
        description:
          $ref: '#/components/schemas/Description'
    AppointmentDetail:
      allOf:
        - $ref: '#/components/schemas/AppointmentBasic'
...
    ClinicBasic:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          minLength: 6
          maxLength: 64
          example: Clinic name
          default: Clinic name
...
    MedicationCreate:
      type: object
      required: [name, value]
      properties:
        name:
          $ref: '#/components/schemas/MedicationType'
        value:
          $ref: '#/components/schemas/Trilian'
        description:
          $ref: '#/components/schemas/Description'
...
    SurgeryCreate:
      type: object
      required: [name, date]
      properties:
        name:
          type: string
          example: "Surgery"
          default: "Surgery"
        date:
          type: string
          format: yyyy-MM-dd
          example: "2018-10-30"
          default: "2018-10-30"
        description:
          $ref: '#/components/schemas/Description'
...
    SymptomCreate:
      type: object
      required: [name, note]
      properties:
        name:
          type: string
          example: "Head"
          default: "Head"
        note:
          type: string
          example: "Headache"
          default: "Headache"
        description:
          $ref: '#/components/schemas/Description'
...